+++ src/main/java/com/insane96mcp/shieldsplus/setup/Strings.java
@@ -19,5 +19,8 @@ public class Strings {
     public static class Translatable {
         public static final String DAMAGE_BLOCKED = ShieldsPlus.RESOURCE_PREFIX + "damage_blocked";
         public static final String REINFORCED_BONUS = ShieldsPlus.RESOURCE_PREFIX + "reinforced_bonus";
+        public static final String REFLECTION_MALUS = ShieldsPlus.RESOURCE_PREFIX + "reflection_malus";
+        public static final String DAMAGE_REFLECTED = ShieldsPlus.RESOURCE_PREFIX + "damage_reflected";
+        public static final String CAPPED_DAMAGE_REFLECTED = ShieldsPlus.RESOURCE_PREFIX + "capped_damage_reflected";
 	}
 }