+++ changelog.md
@@ -4,7 +4,8 @@
 * Added 3 shields enchantments
   * Reinforced: increases damage blocked by 0.5 per level (max IV)
   * Recoil: Knockbacks entities and launches back projectiles that hit the shield (max II)
-  * Pointed: Damages entities that hit the shield by 1 per level (max II)
+  * Reflection: Damages entities that hit the shield by 15% per level of the damage taken (max III)
+  * Ablaze
 * Shields can now be enchanted at the Enchanting Table
 
 ## 1.1.1